refactor: useImageDownload를 유틸로 분리
